feat: Make the Fluent API the only public API and fix cache correctness
The low-level support traits move to a private module, so db.fluent() is now the
only supported entry point. Cached list/query require a preloaded collection:
read-through falls back to Firestore, cached-only errors instead of silently
returning partial results. Also fixes the persistent backend never committing
document deletions.
