+++ src/db/get.rs
@@ -14,173 +14,6 @@ use serde::Deserialize;
 use tracing::*;
 
 #[async_trait]
-pub trait FirestoreGetByIdSupport {
-    async fn get_doc<S>(
-        &self,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<Document>
-    where
-        S: AsRef<str> + Send;
-
-    async fn get_doc_at<S>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<Document>
-    where
-        S: AsRef<str> + Send;
-
-    async fn get_obj<T, S>(&self, collection_id: &str, document_id: S) -> FirestoreResult<T>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn get_obj_return_fields<T, S>(
-        &self,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<T>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn get_obj_at<T, S>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_id: S,
-    ) -> FirestoreResult<T>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn get_obj_at_return_fields<T, S>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<T>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn get_obj_if_exists<T, S>(
-        &self,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<Option<T>>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn get_obj_at_if_exists<T, S>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_id: S,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<Option<T>>
-    where
-        for<'de> T: Deserialize<'de>,
-        S: AsRef<str> + Send;
-
-    async fn batch_stream_get_docs<S, I>(
-        &self,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<(String, Option<Document>)>>
-    where
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_docs_with_errors<S, I>(
-        &self,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<FirestoreResult<(String, Option<Document>)>>>
-    where
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_docs_at<S, I>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<(String, Option<Document>)>>
-    where
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_docs_at_with_errors<S, I>(
-        &self,
-        parent: &str,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<FirestoreResult<(String, Option<Document>)>>>
-    where
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_objects<'a, T, S, I>(
-        &'a self,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<'a, (String, Option<T>)>>
-    where
-        for<'de> T: Deserialize<'de> + Send + 'a,
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_objects_with_errors<'a, T, S, I>(
-        &'a self,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<'a, FirestoreResult<(String, Option<T>)>>>
-    where
-        for<'de> T: Deserialize<'de> + Send + 'a,
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_objects_at<'a, T, S, I>(
-        &'a self,
-        parent: &str,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<'a, (String, Option<T>)>>
-    where
-        for<'de> T: Deserialize<'de> + Send + 'a,
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-
-    async fn batch_stream_get_objects_at_with_errors<'a, T, S, I>(
-        &'a self,
-        parent: &str,
-        collection_id: &str,
-        document_ids: I,
-        return_only_fields: Option<Vec<String>>,
-    ) -> FirestoreResult<BoxStream<'a, FirestoreResult<(String, Option<T>)>>>
-    where
-        for<'de> T: Deserialize<'de> + Send + 'a,
-        S: AsRef<str> + Send,
-        I: IntoIterator<Item = S> + Send;
-}
-
-#[async_trait]
 impl FirestoreGetByIdSupport for FirestoreDb {
     async fn get_doc_at<S>(
         &self,
